GLFeatureInfo: workaround gl bug with nouveau driver
Required to have --use-gl=egl working with nouveau driver.
Otherwise there is a side effect on the OFF-screen context.
It will fail to be created because a gl error is pending
on the ON-screen context.

This bug will go away if glTexImage2D succeeds.
Or if glCheckFramebufferStatusEXT properly return false
when associated glTexImage2D failed.