Allow to pass ids of gpu driver workarounds to command line
For example with workaround "use_virtualized_gl_contexts",
currently the only way to enable it from the command line
is to pass --use_virtualized_gl_contexts.
(--use_virtualized_gl_contexts=0 to disable it)

This patch allows to enable it providing its id:
--gpu-driver-bug-workarounds=62

Generally the patch allows to provides a set of id:
--gpu-driver-bug-workarounds=i,j,k,l

These ids are append to the auto generated workarounds.
So maybe another gpu switch should be use for that:
kGpuDriverBugWorkaroundsAdd instead of kGpuDriverBugWorkarounds.
Then it would be possible to do:
--gpu-driver-bug-workarounds-add=+i,+j,-k,-l,+m
It already exists switches that interpret +/- so it is consistent
and one could re-use these parsers.